Project_3/requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#



-i https://pypi.org/simple

altair==4.2.0; python_version >= '3.7'

argon2-cffi-bindings==21.2.0; python_version >= '3.6'

argon2-cffi==21.3.0; python_version >= '3.6'

asttokens==2.0.5

attrs==22.1.0; python_version >= '3.5'

backcall==0.2.0

beautifulsoup4==4.11.1; python_version >= '3.6'

bleach==5.0.1; python_version >= '3.7'

blinker==1.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

blis==0.7.8

cachetools==5.2.0; python_version ~= '3.7'

catalogue==2.0.8; python_version >= '3.6'

certifi==2022.6.15; python_version >= '3.6'

cffi==1.15.1

charset-normalizer==2.1.0; python_version >= '3.6'

click==8.1.3; python_version >= '3.7'

colorama==0.4.5; platform_system == 'Windows'

commonmark==0.9.1

cymem==2.0.6

debugpy==1.6.2; python_version >= '3.7'

decorator==5.1.1; python_version >= '3.5'

defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

entrypoints==0.4; python_version >= '3.6'

executing==0.9.1

fastjsonschema==2.16.1

gitdb==4.0.9; python_version >= '3.6'

gitpython==3.1.27; python_version >= '3.7'

idna==3.3; python_version >= '3.5'

importlib-metadata==4.12.0; python_version >= '3.7'

ipykernel==6.15.1; python_version >= '3.4'

ipython-genutils==0.2.0

ipython==8.4.0; python_version >= '3.8'

ipywidgets==7.7.1

jedi==0.18.1; python_version >= '3.6'

jinja2==3.1.2; python_version >= '3.7'

joblib==1.1.0; python_version >= '3.6'

jsonschema==4.8.0; python_version >= '3.7'

jupyter-client==7.3.4; python_version >= '3.7'

jupyter-core==4.11.1; python_version >= '3.7'

jupyterlab-pygments==0.2.2; python_version >= '3.7'

jupyterlab-widgets==1.1.1; python_version >= '3.6'

langcodes==3.3.0; python_version >= '3.6'

markupsafe==2.1.1; python_version >= '3.7'

matplotlib-inline==0.1.3; python_version >= '3.5'

mistune==0.8.4

murmurhash==1.0.7

nbclient==0.6.6; python_version >= '3.7'

nbconvert==6.5.0; python_version >= '3.7'

nbformat==5.4.0; python_version >= '3.7'

nest-asyncio==1.5.5; python_version >= '3.5'

notebook==6.4.12; python_version >= '3.7'

numpy==1.23.1

openpyxl

packaging==21.3; python_version >= '3.6'

pandas==1.4.3

pandocfilters==1.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

parso==0.8.3; python_version >= '3.6'

pathy==0.6.2; python_version >= '3.6'

pickleshare==0.7.5

pillow==9.2.0; python_version >= '3.7'

preshed==3.0.6

prometheus-client==0.14.1; python_version >= '3.6'

prompt-toolkit==3.0.30; python_full_version >= '3.6.2'

protobuf==3.20.1; python_version >= '3.7'

psutil==5.9.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

pure-eval==0.2.2

pyarrow==8.0.0; python_version >= '3.7'

pycparser==2.21

pydantic==1.9.1; python_full_version >= '3.6.1'

pydeck==0.7.1; python_version >= '3.7'

pygments==2.12.0; python_version >= '3.6'

pympler==1.0.1; python_version >= '3.6'

pyparsing==3.0.9; python_full_version >= '3.6.8'

pyrsistent==0.18.1; python_version >= '3.7'

python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

pytz-deprecation-shim==0.1.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'

pytz==2022.1

pywin32==304; sys_platform == 'win32' and platform_python_implementation != 'PyPy'

pywinpty==2.0.6; os_name == 'nt'

pyzmq==23.2.0; python_version >= '3.6'

requests==2.28.1; python_version >= '3.7' and python_version < '4.0'

rich==12.5.1; python_version < '4.0' and python_full_version >= '3.6.3'

scikit-learn==1.1.1; python_version >= '3.8'

scipy==1.8.1; python_version < '3.11' and python_version >= '3.8'

semver==2.13.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

send2trash==1.8.0

setuptools==63.2.0; python_version >= '3.7'

six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

sklearn==0.0

smart-open==5.2.1; python_version >= '3.6' and python_version < '4.0'

smmap==5.0.0; python_version >= '3.6'

soupsieve==2.3.2.post1; python_version >= '3.6'

spacy-legacy==3.0.9; python_version >= '3.6'

spacy-loggers==1.0.3; python_version >= '3.6'

spacy==3.4.1

srsly==2.4.4; python_version >= '3.6'

stack-data==0.3.0

streamlit==1.11.1

terminado==0.15.0; python_version >= '3.7'

thinc==8.1.0; python_version >= '3.6'

threadpoolctl==3.1.0; python_version >= '3.6'

tinycss2==1.1.1; python_version >= '3.6'

toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'

toolz==0.12.0; python_version >= '3.5'

tornado==6.2; python_version >= '3.7'

tqdm==4.64.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

traitlets==5.3.0; python_version >= '3.7'

typer==0.4.2; python_version >= '3.6'

typing-extensions==4.3.0; python_version >= '3.7'

tzdata==2022.1; platform_system == 'Windows'

tzlocal==4.2; python_version >= '3.6'

urllib3==1.26.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4.0'

validators==0.20.0; python_version >= '3.4'

wasabi==0.10.1

watchdog==2.1.9; platform_system != 'Darwin'

wcwidth==0.2.5

webencodings==0.5.1

widgetsnbextension==3.6.1

zipp==3.8.1; python_version >= '3.7'
